update: 优化关联相关接口

+++ src/router/controller/roles-permissions/delete.ts
@@ -12,7 +12,8 @@ import { query } from "../../../db";
  * 角色-权限关联删除
 */
 export const doRolePermissionDelete = async (ctx: Context, next: Next) => {
-  const sql: string = `DELETE FROM roles_permissions WHERE id = ?`
-  await query(sql, ctx._params.id)
+  const sql: string = `DELETE FROM roles_permissions WHERE id = ? OR (role_id = ? AND permission_id = ?)`
+  const data = [ctx._params.id, ctx._params.roleId, ctx._params.permissionId]
+	await query(sql, data)
   throw new Success();
 }
